Enable todo delete and mark as done

+++ src/main/java/com/bbh/repository/TodoActivityRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bbh.model.TodoActivity;
 
-public interface TodoActivityRepo extends JpaRepository<TodoActivity, Long> {
+public interface TodoActivityRepo extends JpaRepository<TodoActivity, Integer> {
 
 }
